Add files via upload
As COVID-19 continues to spread, Covid tracker and Vaccination app help to contain the pandemic. The main objective of this application is to form it interactive and its ease to use where you'll be able to check what number people are affected and how many already died and recovered with the assistance of graph statistics. It also contains a menu shows the symptoms of COVID-19 and prevention to stay people safe. App displays available vaccination centers and necessary information regarding center name, location, vaccine name, availability of dose1 and dose 2 and other related information. When it comes to authentication it uses firebase authentication. It verifies email implemented through firebase. The app allows user to require self-assessment test and lets them know their risk level.